backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-jose
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy
aiosqlite
requests
fpdf2
python-dotenv
python-nmap
slowapi
